fixed-profile#removed drawings number

+++ initializeDB.js
@@ -57,9 +57,10 @@ function dbrun() {
         if (err) throw err;
 
         for (const file of files) {
-            fs.unlink(path.join(directory, file), (err) => {
-                if (err) throw err;
-            });
+            if (file != 'default.png')
+                fs.unlink(path.join(directory, file), (err) => {
+                    if (err) throw err;
+                });
         }
     });
 }
